requirements.txt
streamlit
psutil
matplotlib
groq
python-dotenv